chore: useSelectHook removed
