+++ src/Shared/Components/SelectPicker/constants.ts
@@ -17,17 +17,6 @@
 import { CSSProperties } from 'react'
 import { ComponentSizeType } from '@Shared/constants'
 import { SelectPickerProps } from './type'
-import {
-    SelectPickerLoadingIndicator,
-    SelectPickerDropdownIndicator,
-    SelectPickerControl,
-    SelectPickerOption,
-    SelectPickerMenuList,
-    SelectPickerClearIndicator,
-    SelectPickerMultiValueRemove,
-    SelectPickerInput,
-    SelectPickerValueContainer,
-} from './common'
 
 export const SELECT_PICKER_FONT_SIZE_MAP: Record<SelectPickerProps['size'], CSSProperties['fontSize']> = {
     [ComponentSizeType.small]: '12px',
@@ -48,17 +37,4 @@ export const SELECT_PICKER_CONTROL_SIZE_MAP: Record<SelectPickerProps['size'], C
     [ComponentSizeType.medium]: 'auto',
     [ComponentSizeType.large]: '36px',
     [ComponentSizeType.xl]: '36px',
-}
-
-export const BaseSelectComponents = {
-    IndicatorSeparator: null,
-    LoadingIndicator: SelectPickerLoadingIndicator,
-    DropdownIndicator: SelectPickerDropdownIndicator,
-    Control: SelectPickerControl,
-    Option: SelectPickerOption,
-    MenuList: SelectPickerMenuList,
-    ClearIndicator: SelectPickerClearIndicator,
-    MultiValueRemove: SelectPickerMultiValueRemove,
-    Input: SelectPickerInput,
-    ValueContainer: SelectPickerValueContainer,
 }
